autofill $version array for update process
Before each release, you have to edit a static property of
the Update class, and this is pretty complex when this has to match
several branch version (for example 2.0.x/2.1.x/2.2.x).

This modification remove the static and use the content of SQL_DIR
to fill the array Update->version (non static) with all version.

With that, your release process will not require anymore to update that
file.

Last point: if you use the update process from the master branch, you
will update to a «not released yet version», which can still change
later), but as it's a development branch, so this should not be a problem.
Add to that this provide an easy way to test the update process before
a release and without having to edit anything.

Conflicts:
	core/lib/Thelia/Install/Update.php