Fix attributes and path commerce guys address (#3226)
* Update SchemaLocator.php

Replaces the use of a relative path with an absolute path. For example, when the generation of models in the cache is called from the Thelia php console, it uses the path local/config/shema.xml. Whereas if the generation of models in the cache is called from an http request, it uses the path web/local/config/shema.xml.

* Update SchemaLocator.php

fix code style

* fix the path of commerceguys address and its attributes

---------

Co-authored-by: hequet <erwann.hequet@api-and-you.com>